Upload a directory's contents

Upload a directory as a new versioned asset of a project in the registry.

uploadDirectory(
  project,
  asset,
  version,
  directory,
  staging,
  prefix = NULL,
  probation = FALSE,
  owners = NULL,
  uploaders = NULL
)

Arguments

project

String containing the project name. This may be NULL, in which case prefix should be provided.

asset

String containing the asset name.

version

String containing the version name. This may be NULL to generate a new version name based on an incrementing series.

directory

String containing the path to a directory to be uploaded. This should be inside staging for best performance, see allocateUploadDirectory for details.

staging

String containing the path to the staging directory.

prefix

String containing the project name prefix. This is used to derive a project name based on an incrementing series.

probation

Logical scalar indicating whether to upload a probational version.

owners

Character vector containing the user IDs for owners of this project. This defaults to the current user.

uploaders

List specifying the authorized uploaders for this project. See the uploaders field in the fetchPermissions return value for the expected format.

Value

List containing project and version, the generated names for the project and version respectively. (If these were explicitly provided in the arguments, the same values are returned verbatim.)

See also

Author

Aaron Lun

Examples

info <- startGobbler()

src <- allocateUploadDirectory(info$staging)
write(file=file.path(src, "foo"), "BAR")

res <- uploadDirectory(
    project=NULL, 
    prefix="TEST", 
    asset="simple", 
    version=NULL, 
    directory=src, 
    staging=info$staging)
res
#> $project
#> [1] "TEST1"
#> 
#> $status
#> [1] "SUCCESS"
#> 
#> $version
#> [1] "1"
#>